Implement `foreign_key_optional` for `HasMany`
This PR adds a new attribute to `#[has_many]` that enables support for
HasMany relations where the child foreign key is optional.

This PR is mainly plumbing for getting the attribute through to
`child_ids_impl` and then a small bit for using the value to decide
whether to wrap the parent id in `Some(..)`.

Fixes #7