Allow use of mixed id types.
I'm not sure it was ever intended to not work, but a few places in the
codebase depended on the id of a child being the same type as the id of
a parent, which certainly cannot be expected to always hold.

This commit solves the issue by referencing the childs actual id using
the Id associated type from GraphqlNodeForModel.